SmoothDamp reformatting
Used Vector3s to define variables instead of floats, reduced the number of lines of code and combined some multiplication and division operations in order to reduce the overall count. Perhaps it still works the same and gets the same results. There could be some rounding difference in compare to the previous version.